Fix double ended queue did not print free strings
Because it is possible for to_string function to generate a new string
here and nothing else sees it after printing it is not guaranteed to be
freed leading to a memory leak. This does mean that if the string is
internal to a struct and expected to be used later then to_string should
duplicate said string (as noted in the documentation).